requirements.txt
selenium
webdriver_manager
flask-restful